# server/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API COLECTIVO HIFAS DE GRANA",
        "description": "Servicios para la página web del colectivo Hifas de Grana, la cual se encarga de crear conciencia difundiendo y realizando actividades que generen reflexión hacia el cuidado del medio ambiente.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http"
    ],
    "consumes":[
        "application/json"
    ],
    "produces" :[
        "application/json"
    ],
    "paths":
...
        "/api/contacto/mostrarContactos":
        {
            "get":
            {
                "summary": "Muestra a los contactos",
                "operationId": "mostrarContactos",
                "produces": ["application/json"],
                "tags": ["conecta"],
                "responses": {
                    "200": {
                        "description": "contactos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/contacto"
                            }
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "name": "Juan",
                                    "email": "juan.perez@example.com",
                                    "message": "Hola, este es un mensaje de prueba."
                                },
                                {
                                    "id": 2,
                                    "name": "Maria",
                                    "email": "maria.gomez@example.com",
                                    "message": "Hola, este es otro mensaje de prueba."
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error al mostrar contactos"
                    }
                },
                "description": "Muestra a aquellos ususarios interesados en contactar con el colectivo"
            }
        },
...
        "/api/contacto/actualizarContacto/{id}": {
            "put": {
                "summary": "Actualiza un contacto",
                "operationId": "actualizarContacto",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del contacto",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "El objeto nuevoContacto que se va a actualizar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nuevoContacto"
                        }
                    }
                ],
                "tags": ["conecta"],
                "responses": {
                    "200": {
                        "description": "contacto actualizado",
                        "schema": {
                            "$ref": "#/definitions/contacto"
                        }
                    },
                    "400": {
                        "description": "Error al actualizar contacto"
                    }
                },
                "description": "Actualiza un contacto"
            }
        },
        "/api/contacto/buscarContacto/{id}":
        {
            "get":
            {
                "summary": "Busca un contacto",
                "operationId": "buscarContacto",
                "parameters" : [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del contacto",
                        "required": true,
                        "shema": {
                            "$ref": "#/definitions/contacto"
                        }
                    }
                ],
                "tags": ["conecta"],
                "responses": {
                    "200": {
                        "description": "contacto encontrado",
                        "schema": {
                            "$ref": "#/definitions/contacto"
                        }
                    },
                    "400": {
                        "description": "Error al buscar contacto"
                    }
                },
                "description": "Busca un contacto"
            }
        },
...
        "/api/recursos_educativos/mostrarRecursos_educativos":
        {
            "get":
            {
                "summary": "Muestra los recursos educativos",
                "operationId": "mostrarRecursos_educativos",
                "produces": ["application/json"],
                "tags": ["recursos_educativos"],
                "responses": {
                    "200": {
                        "description": "recursos educativos",
                        "schema": {
                            "$ref": "#/definitions/recurso_educativo"
                        }
                    },
                    "400": {
                        "description": "Error al mostrar recursos educativos"
                    }
                },
                "description": "Muestra los recursos educativos"
            }
        },
...
        "/api/recursos_educativos/actualizarRecurso_educativo/{id}":
        {
            "put":
            {
                "summary": "Actualiza un recurso educativo",
                "operationId": "actualizarRecurso_educativo",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters" : [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del recurso educativo",
                        "required": true,
                        "shema": {
                            "$ref": "#/definitions/recurso_educativo"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "El objeto nuevoRecurso_educativo que se va a actualizar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nuevoRecurso_educativo"
                        }
                    }
                ],
                "tags": ["recursos_educativos"],
                "responses": {
                    "200": {
                        "description": "recurso educativo actualizado",
                        "schema": {
                            "$ref": "#/definitions/recurso_educativo"
                        }
                    },
                    "400": {
                        "description": "Error al actualizar recurso educativo"
                    }
                },
                "description": "Actualiza un recurso educativo"
            }
        },
...
        "/api/donativos/mostrarDonativos":
        {
            "get":
            {
                "summary": "Muestra los donativos",
                "operationId": "mostrarDonativos",
                "produces": ["application/json"],
                "tags": ["donativos"],
                "responses": {
                    "200": {
                        "description": "donativos",
                        "schema": {
                            "$ref": "#/definitions/donativo"
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "nombre": "Donativo 1",
                                    "cantidad": 100,
                                    "fecha": "2022-01-01"
                                },
                                {
                                    "id": 2,
                                    "nombre": "Donativo 2",
                                    "cantidad": 200,
                                    "fecha": "2022-01-02"
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error al mostrar donativos"
                    }
                },
                "description": "Muestra los donativos"
            }
        },        
...
        "/api/donativos/actualizarDonativo/{id}":
        {
            "put":
            {
                "summary": "Actualiza un donativo",
                "operationId": "actualizarDonativo",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters" : [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del donativo",
                        "required": true,
                        "shema": {
                            "$ref": "#/definitions/donativo"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "El objeto nuevoDonativo que se va a actualizar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nuevoDonativo"
                        }
                    }
                ],
                "tags": ["donativos"],
                "responses": {
                    "200": {
                        "description": "donativo actualizado",
                        "schema": {
                            "$ref": "#/definitions/donativo"
                        }
                    },
                    "400": {
                        "description": "Error al actualizar donativo"
                    }
                },
                "description": "Actualiza un donativo"
            }
        }      
    },
    "definitions":{
...
        "contacto" : {
            "type": "object",
            "propierties" : {
                "id" : {
                    "type": "integer",
                    "description": "id del contacto"
                },
                "name" : {
                    "type": "string",
                    "description": "nombre del contacto"
                },
                "email" : {
                    "type": "string",
                    "description": "email del contacto"
                },
                "mensaje" : {
                    "type": "string",
                    "description": "mensaje del contacto"
                }
            },
            "example": {
                "application/json": {
                    "id": 1,
                    "name": "Juan",
                    "email": "juan.perez.actualizado@example.com",
                    "message": "Hola, este es un mensaje de prueba actualizado."
                }
            }
        },
...
        "nuevoContacto" : {
            "type": "object",
            "properties" : {
                "name" : {
                    "type": "string",
                    "description": "nombre del contacto"
                },
                "email" : {
                    "type": "string",
                    "description": "email del contacto"
                },
                "message" : {
                    "type": "string",
                    "description": "mensaje del contacto"
                }
            }
        },
        "recurso_educativo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "tipo": {
                    "type": "string"
                },
                "archivo": {
                    "type": "string"
                },
                "autor": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                }
            },
            "example": {
                "id": 1,
                "tipo": "Tipo 1",
                "archivo": "Archivo 1",
                "autor": "Autor 1",
                "titulo": "Titulo 1"
            }
        },
...
        "nuevoRecurso_educativo": {
            "type": "object",
            "properties": {
                "tipo": {
                    "type": "string"
                },
                "archivo": {
                    "type": "string"
                },
                "autor": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },        
        "donativo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "direccion": {
                    "type": "string"
                },
                "correo_electronico": {
                    "type": "string"
                },
                "numero_telefono": {
                    "type": "string"
                },
                "fecha_donacion": {
                    "type": "string",
                    "format": "date-time"
                },
                "cantidad_donada": {
                    "type": "number",
                    "format": "double"
                },
                "estado_donacion": {
                    "type": "string"
                }
            },
            "example": {
                "id": 1,
                "direccion": "123 Calle Principal",
                "correo_electronico": "donante@example.com",
                "numero_telefono": "1234567890",
                "fecha_donacion": "2022-01-01T00:00:00Z",
                "cantidad_donada": 100.0,
                "estado_donacion": "Aceptado"
            }
        },
...
        "nuevoDonativo": {
            "type": "object",
            "properties": {
                "direccion": {
                    "type": "string"
                },
                "correo_electronico": {
                    "type": "string"
                },
                "numero_telefono": {
                    "type": "string"
                },
                "fecha_donacion": {
                    "type": "string",
                    "format": "date-time"
                },
                "cantidad_donada": {
                    "type": "number",
                    "format": "double"
                },
                "estado_donacion": {
                    "type": "string"
                }
            }
        }
    }
}